+++ src/string/mod.rs
@@ -24,7 +24,7 @@ pub struct StringLit<B: Buffer> {
 
     /// The number of hash signs in case of a raw string literal, or `None` if
     /// it's not a raw string literal.
-    num_hashes: Option<u32>,
+    num_hashes: Option<u8>,
 
     /// Start index of the suffix or `raw.len()` if there is no suffix.
     start_suffix: usize,
@@ -110,7 +110,7 @@ impl<B: Buffer> fmt::Display for StringLit<B> {
 
 /// Precondition: input has to start with either `"` or `r`.
 #[inline(never)]
-pub(crate) fn parse_impl(input: &str) -> Result<(Option<String>, Option<u32>, usize), ParseError> {
+pub(crate) fn parse_impl(input: &str) -> Result<(Option<String>, Option<u8>, usize), ParseError> {
     if input.starts_with('r') {
         scan_raw_string(&input, 1, true, true)
             .map(|(hashes, start_suffix)| (None, Some(hashes), start_suffix))
